Improve code structure
Move all statment related logic to `statement.rs` and expression related logic to `expression.rs`.
